# backend/swagger/swagger.yaml
...
openapi: 3.1.1
info:
  title: KeyAuthority
  version: 1.4.35
  description: |
    Welcome to the **KeyAuthority REST API** documentation.

    KeyAuthority is a production-grade PKI and secret management system designed 
    for seamless integration with Kubernetes, cloud-native environments, and modern security architectures. 
    This documentation provides a complete reference for most endpoints, 
    including PKI operations, secret management, and health checks.

    We document here only the status error code 400 for simplicity. In practice, the API may return other status codes such as 428 for actions requiring approval, 401 for unauthorized access, 403 for forbidden actions, and 500 for internal server errors. The error response will include a JSON body with an "errors" field describing the error messages.

    **Features:**
    - Full support for X.509 CA operations
    - Native integration with [cert-manager](https://cert-manager.io)
    - Secret storage and decryption compatible with widely used tools like HashiCorp Vault
    - HSM-backed signing and key management

    **Disclaimer:** This product partially implements a HashiCorp Vault-compatible API. HashiCorp Vault, Vault Agent (injector), and the `vault` CLI are trademarks of HashiCorp, Inc. This project is not affiliated with or endorsed by HashiCorp.
servers:
  - url: /
...
paths:
...
  /v1/keys:
    get:
      summary: Get keys
      description: |
        Get the accessible keys with optional filtering and pagination.
      tags:
        - Keys
      parameters:
        - name: id
          in: query
          schema:
            type: string
            format: uuid
        - name: type
          in: query
          schema:
            type: string
        - name: storage
          in: query
          schema:
            type: string
            enum: [Software, HSM]
        - name: environment
          in: query
          schema:
            type: string
        - name: cursor
          in: query
          schema:
            type: string
        - name: pageSize
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Key"
                  nextCursor:
                    type: string
                  pageSize:
                    type: integer
                  hasMore:
                    type: boolean
        "400":
          description: Failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security:
        - bearerAuth: []
    post:
      summary: Create key
      description: |
        Create a new key with the specified configuration.
      tags:
        - Keys
      parameters:
        - name: environment
          example: production
          required: true
          in: query
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KeyConfig"
      responses:
        "201":
          description: Success
          content:
            text/plain:
              schema:
                type: string
                format: uuid
        "400":
          description: Failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security:
        - bearerAuth: []
...
components:
  schemas:
    KeyConfig:
      type: object
      required: [type]
      properties:
        type:
          type: string
          enum: [RSA, ECDSA, Ed25519, AES]
          description: Type of key
        bits:
          type: integer
          # example: 2048
          description: Key size in bits
        curve:
          type: string
          enum: [P-224, P-256, P-384, P-521]
          description: Elliptic curve for ECDSA keys
        mode:
          type: string
          enum: [GCM]
          description: Cipher mode for AES keys
        pkcs11URI:
          type: string
          description: PKCS#11 URI for HSM keys
        pkcs11KeyURI:
          type: string
          description: PKCS#11 key URI for HSM keys
      example:
        type: RSA
        bits: 2048
        pkcs11URI: pkcs11:module-path=/usr/lib64/pkcs11/libsofthsm2.so;token=keyauthority?pin-source=/etc/softhsm/.pin
        pkcs11KeyURI: pkcs11:id=17629ea8924cf716be8e97da23ab621b;object=keyauthority
    Key:
      type: object
      properties:
        id:
          type: string
          format: uuid
        environment:
          type: string
        createdAt:
          type: string
          format: date-time
        config:
          $ref: "#/components/schemas/KeyConfig"
...
    Error:
      type: object
      properties:
        errors:
          type: array
          items:
            type: string
          description: Error messages
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication